old/requirements.txt
# langchain core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic


# Google Generative AI Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Management
python-dotenv

# Machine Learning Utilities
scikit-learn
numpy
